compiler/natives/src/math/big: Remove bitLen, override Word.
math/big package no longer has a bitLen low-level function,
so it can be removed.

In Go 1.9, the type of big.Word has changed from uintptr
to uint. Due to issue #652 with uint multiplication with
overflows, we use a workaround of overriding type of Word
back to uintptr. This should be undone once #652 is fixed.

Fixes:

	$ gopherjs test --short math/big
	$GOROOT/src/math/big/big.go:39:9: undeclared name: bitLen_g